# Compare `*stacktraceTree.insert` against different initial sizes